test: integrationTest 분할 Related to: #66
